Move conda.ts
