+++ src/test/pythonEnvironments/discovery/locators/condaLocator.unit.test.ts
@@ -6,7 +6,7 @@ import * as path from 'path';
 import * as sinon from 'sinon';
 import { PythonReleaseLevel, PythonVersion } from '../../../../client/pythonEnvironments/base/info';
 import * as externalDeps from '../../../../client/pythonEnvironments/common/externalDependencies';
-import { getPythonVersionFromConda } from '../../../../client/pythonEnvironments/discovery/locators/services/conda';
+import { getPythonVersionFromConda } from '../../../../client/pythonEnvironments/common/environmentManagers/conda';
 import { TEST_DATA_ROOT } from '../../common/commonTestConstants';
 import { assertVersionsEqual } from './envTestUtils';
